Upgrading react-native-0.29.0
