+++ android/app/src/main/java/com/transistorsoft/backgroundgeolocation/react/MainActivity.java
@@ -1,0 +1,15 @@
+package com.transistorsoft.backgroundgeolocation.react;
+
+import com.facebook.react.ReactActivity;
+
+public class MainActivity extends ReactActivity {
+
+    /**
+     * Returns the name of the main component registered from JavaScript.
+     * This is used to schedule rendering of the component.
+     */
+    @Override
+    protected String getMainComponentName() {
+        return "RNBackgroundGeolocationSample";
+    }
+}
