+++ android/settings.gradle
@@ -8,9 +8,6 @@ project(':react-native-background-geolocation').projectDir = new File(rootProjec
 include ':react-native-vector-icons'
 project(':react-native-vector-icons').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-vector-icons/android')
 
-include ':react-native-gmaps'
-project(':react-native-gmaps').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-gmaps/android/app')
-
 include ':RNDeviceInfo'
 project(':RNDeviceInfo').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-device-info/android')
