+++ package.json
@@ -1,18 +1,18 @@
 {
-  "name": "RNBackgroundGeolocationSample",
+  "name": "BackgroundGeolocationDemo",
   "version": "2.0.0",
   "private": true,
   "scripts": {
     "start": "node_modules/react-native/packager/packager.sh"
   },
   "dependencies": {
-    "react-native": "^0.25.1",
-    "react": "^0.14.5",
+    "react-native": "^0.29.2",
+    "react": "^15.2.1",
     "react-native-background-geolocation": "^1.1.0",
-    "react-native-device-info": "^0.8.3",
-    "react-native-drawer": "^2.2.2",
-    "react-native-mapbox-gl": "^4.0.0",
-    "react-native-modalbox": "^1.3.3",
-    "react-native-vector-icons": "^2.0.2"
+    "react-native-device-info": "^0.9.3",
+    "react-native-drawer": "^2.2.6",
+    "react-native-mapbox-gl": "^4.1.1",
+    "react-native-modalbox": "^1.3.4",
+    "react-native-vector-icons": "^2.0.3"
   }
 }
